move packet types to types v2
